requirements.txt
streamlit
duckdb
pandas
numpy
altair
plotly
google-generativeai
python-dotenv
